+++ app/assets/javascripts/common/image_click.js
