+++ app/assets/stylesheets/show.scss
@@ -22,7 +22,7 @@
       font-weight: bold;
     }
 
-    &_intro {
+    &_intro , &_age {
       margin-top: 30px;
       text-align: center;
     }
