+++ app/controllers/favorites_controller.rb
@@ -7,6 +7,7 @@ class FavoritesController < ApplicationController
     favorite.favorite_count(@user) if favorite.favorite?(params[:favorite][:status])
     favorite.save
     @favorite = Favorite.find_by(to_user_id: @user.id, from_user_id: current_user.id)
+    room_create(@user) if matching?(@user)
   end
 
   def following
@@ -27,5 +28,19 @@ class FavoritesController < ApplicationController
 
   def create_params
     params.require(:favorite).permit(:to_user_id, :from_user_id, :status)
+  end
+
+  def matching?(user)
+    @matching = Favorite.find_by(to_user_id: current_user.id, from_user_id: user.id, status: 1)
+  end
+
+  def room_create(user)
+    user_rooms = RoomUser.where(user_id: current_user.id).map(&:room)
+    chat_room = RoomUser.where(room: user_rooms, user_id: user.id).map(&:room).first
+    if chat_room.blank?
+      chat_room = Room.create
+      RoomUser.create(room: chat_room, user_id: current_user.id)
+      RoomUser.create(room: chat_room, user_id: user.id)
+    end
   end
 end
