+++ app/controllers/matching_controller.rb
@@ -2,7 +2,8 @@ class MatchingController < ApplicationController
   before_action :authenticate_user!
 
   def index
-    favorite_user_ids = Favorite.where(to_user_id: current_user.id, status: 1).pluck(:from_user_id)
-    @matching = Favorite.where(to_user_id: favorite_user_ids, from_user_id: current_user.id, status: 1).map(&:to_user)
+    @rooms = current_user.rooms
+    room_users_ids = RoomUser.where(room_id: @rooms).where.not(user_id: current_user.id).pluck(:user_id)
+    @room_users = User.where(id: room_users_ids)
   end
 end
